+++ modules/apps/asset/asset-link-service/src/main/java/com/liferay/asset/link/service/impl/AssetLinkLocalServiceImpl.java
@@ -24,12 +24,12 @@ import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.mass.delete.MassDeleteCacheThreadLocal;
 import com.liferay.portal.kernel.model.SystemEventConstants;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.SystemEventLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.BulkDeleteCacheThreadLocal;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.model.adapter.util.ModelAdapterUtil;
@@ -188,7 +188,7 @@ public class AssetLinkLocalServiceImpl extends AssetLinkLocalServiceBaseImpl {
 	@Override
 	public void deleteLinks(long entryId) {
 		Map<Long, List<AssetLink>> leftPartitionAssetLinks =
-			BulkDeleteCacheThreadLocal.getBulkDeleteCache(
+			MassDeleteCacheThreadLocal.getBulkDeleteCache(
 				StringBundler.concat(
 					AssetLinkLocalServiceImpl.class.getName(),
 					".deleteLinks#left#", entryId),
@@ -225,7 +225,7 @@ public class AssetLinkLocalServiceImpl extends AssetLinkLocalServiceBaseImpl {
 			});
 
 		Map<Long, List<AssetLink>> rightPartitionAssetLinks =
-			BulkDeleteCacheThreadLocal.getBulkDeleteCache(
+			MassDeleteCacheThreadLocal.getBulkDeleteCache(
 				StringBundler.concat(
 					AssetLinkLocalServiceImpl.class.getName(),
 					".deleteLinks#right#", entryId),
