+++ modules/apps/saved-content/saved-content-service/src/main/java/com/liferay/saved/content/service/impl/SavedContentEntryLocalServiceImpl.java
@@ -8,6 +8,7 @@ package com.liferay.saved.content.service.impl;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.aop.AopService;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.mass.delete.MassDeleteCacheThreadLocal;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.search.Indexable;
 import com.liferay.portal.kernel.search.IndexableType;
@@ -15,7 +16,6 @@ import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.ResourceLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.util.BulkDeleteCacheThreadLocal;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.saved.content.exception.DuplicateSavedContentEntryException;
@@ -100,7 +100,7 @@ public class SavedContentEntryLocalServiceImpl
 		long groupId, long classNameId, long classPK) {
 
 		Map<Long, List<SavedContentEntry>> partitionSavedContentEntries =
-			BulkDeleteCacheThreadLocal.getBulkDeleteCache(
+			MassDeleteCacheThreadLocal.getBulkDeleteCache(
 				StringBundler.concat(
 					SavedContentEntryLocalServiceImpl.class.getName(),
 					".deleteSavedContentEntries#", groupId, classNameId),
